Doc_analyzer - 1/requirements.txt
# Core Libraries
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.51
langchain-text-splitters==0.0.1

# Web scraping
beautifulsoup4==4.12.3
httpx==0.27.0
textstat==0.7.3
undetected-chromedriver==3.5.5
playwright==1.44.0
selenium==4.21.0

# Vector Store and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.6.1
huggingface-hub==0.23.0

# HTML Parsing / Utilities
pydantic==1.10.15
python-dotenv==1.0.1

# Async support
aiohttp==3.9.5

# Required by Playwright (for headless Chromium)
# This installs the browsers (run separately after install)
playwright install

# Optional: for fast JSON handling
ujson==5.10.0
